tests/wpt/tests/secure-payment-confirmation/authentication-invalid-payment-entity-logo.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test for the 'secure-payment-confirmation' payment method authentication - invalid payment entity logo</title>
<link rel="help" href="https://w3c.github.io/secure-payment-confirmation#sctn-steps-to-check-if-a-payment-can-be-made">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="utils.sub.js"></script>
<script>
'use strict';

promise_test(async t => {
  const authenticator = await window.test_driver.add_virtual_authenticator(
    AUTHENTICATOR_OPTS);
  t.add_cleanup(() => {
    return window.test_driver.remove_virtual_authenticator(authenticator);
  });

  await window.test_driver.set_spc_transaction_mode("autoAccept");
  t.add_cleanup(() => {
    return window.test_driver.set_spc_transaction_mode("none");
  });

  const credential = await createCredential();

  const challenge = 'server challenge';
  const payeeOrigin = 'https://merchant.com';
  const displayName = 'Troycard ***1234';

  const paymentEntity1Label = 'Payment Entity #1';
  const paymentEntity2Label = 'Payment Entity #2';

  let request = new PaymentRequest([{
    supportedMethods: 'secure-payment-confirmation',
    data: {
      credentialIds: [credential.rawId],
      challenge: Uint8Array.from(challenge, c => c.charCodeAt(0)),
      payeeOrigin,
      rpId: window.location.hostname,
      timeout: 60000,
      instrument: {
        displayName,
        icon: ICON_URL,
      },
      paymentEntitiesLogos: [
        {
          url: NONEXISTENT_PAYMENT_ENTITY_LOGO_URL,
          label: paymentEntity1Label,
        },
        {
          url: PAYMENT_ENTITY_LOGO_URL,
          label: paymentEntity2Label,
        },
      ],
    }
  }], PAYMENT_DETAILS);

  await test_driver.bless('user activation');
  const responsePromise = request.show();
  const response = await responsePromise;
  await response.complete('success');
  const cred = response.details;
  const clientDataJSON = JSON.parse(arrayBufferToString(cred.response.clientDataJSON));

  // Both PaymentEntityLogos should still be present, but the non-existent url
  // should have been replaced by an empty string.
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos.length, 2);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[0].url, '');
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[0].label, paymentEntity1Label);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[1].url, PAYMENT_ENTITY_LOGO_URL);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[1].label, paymentEntity2Label);
}, 'SPC authentication with a PaymentEntityLogo that cannot be downloaded.');
</script>
